+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This change log follows the conventions of
 - Functions to support drawing bit-fields from integers.
 - Functions to support drawing padding until a particular address is
   reached.
+- A `draw-inline-gap` function to support drawing discontinuities in
+  single-row diagrams.
 
 
 ## [1.3.0] - 2020-03-29